// src/shared/components/profile-details/type.tsx
export type ApiProps = {
    businessName: any;
    businessAddress: any;
    businessEmailAddress: any;
    phoneNumber: any;
    stateId: any;
    city: any;
    zipcode: any;
    businessCategoryType: any;
    businessCategoryTypeId: any;
    profileImage: any,
    businessId: any,
    imageId:any,
    actionPerformedBy:any,
   
  };